tests/krate/publish: Extract `max_size` module
